Managed logging in and logging out in React app, updated Migrations folder

+++ RatingTimeBackEnd/RatingTime.DTO/Models/Movies/MovieDTOs.cs
@@ -2,7 +2,7 @@
 
 namespace RatingTime.DTO.Models.Movies
 {
-    public record MovieInfo(int Id, string Title, string? ImageUrl, string? Overview, double AverageRating);
-    public record MovieDetailInfo(int Id, string Title, string? ImageUrl, string? Overview, double AverageRating, List<GenreInfo>? Genres);
-    public record RatingMoviePost(int Id, string? Title, string? ImageUrl, string? Overview, double AverageRating, List<MovieGenrePost>? Genres);
+    public record MovieInfo(int Id, string Title, string? ImageUrl, string? Overview, decimal AverageRating);
+    public record MovieDetailInfo(int Id, string Title, string? ImageUrl, string? Overview, decimal AverageRating, List<GenreInfo>? Genres);
+    public record RatingMoviePost(int Id, string? Title, string? ImageUrl, string? Overview, decimal AverageRating, List<MovieGenrePost>? Genres);
 }
